Fixed bugs with how GUI data was stored
By default, handles stores pointers to values, not values
themselves.  New convention is to have the pointer names start
with capital letters, then store the corresponding data in handles
with lower case letters. Ex: handles.Metadata for the handle to
the GUI object and handles.metadata to store the table and pass
it to Main_PCO...