Image class nows deals with paths much better
Image.m now has image.dir attribute which is chosen inteligently
when the instance is created with a file_name that includes a
relative path.
Also added load_image.m which automatically sets image.dir when
the object is loaded, which is important if the data has been moved
or is being accessed over a network drive.